Remember previous terminal position and mode
This change keeps the last position of the terminal rather than
automatically scrolling to the bottom each time. This is useful if
compiling the project and wanting to toggle back and forth to see the
error or if there's something in the terminal you want to reference
occasionally.

Also to ensure persist mode works properly I instead call the toggleterm
commands directly rather than changing to normal mode first.