Replace context plugin
The previous context plugin was nice but it was missing a few features
that this plugin provides. Furthermore it conflicted with minimap in a
way that prevented me from using the editor at all. Some new features
include syntax highlighting within the context window as well as showing
the full function along with parameters in the case of a multiline
parameter list.