Add hack to fix neorg fold behaviors
This is (hopefully) a temporary keymap that toggles on and off the
Treesitter playground. I found that toggling the playground fixes the
folding behavior for my notes for neorg. This will be here until an
upstream fix to neorg or whatever repository is necessary to get folds
to behave consistently.