Plugins: Lazy nvim: Turn off partial clones
This change turns off partial clones. During a plane ride I wanted to
look at the blame for a plugin and was unable to because I didn't have
an internet connection. I don't mind the additional storage overhead if
it means I can actually have full history even when offline.