Plugins: Neorg: Add persistent journal keymap
I found that I would commonly open up the journal as one of the first
things that I wanted to do after opening vim. But in order to open it
I'd have to first be in a .norg file so the keymaps would map. So I
would open a file I had no interest in, in order to get the keymaps I
wanted. I decided to just make it a persistent keymap instead.