Plugins: Neorg: Add journal year keymap
I am unsure if there is plans to add various timescales to the journal
module in Neorg. If the people maintaining it didn't feel like writing
it but were okay with it being added I'd be willing to add it. This is
something to hold me over in the mean time. I'll have to ask what the
plans are if any to support year based journal entries. This will have
to be manually updated each year.