Plugins: Update smooth scrolling plugin
This updates the smooth scrolling plugin from using neoscroll to using
mini's animate. I like that the animations are set for more than ctrl-d
and ctrl-u. I wanted smooth scrolling animations for gg and G and this
also implements them among other things.

There is one issue however when using lightspeed. If using lightspeed
and it triggers a vertical scroll due to scrolloff being turned on then
lightspeed has to wait for the animation to finish in order to actually
reach the cursor position. If the mini.animate developer says this can't
be fixed then I may switch back to neoscroll.