Plugins: nvim-treesitter: Allow vim syntax regex
With this disabled if you do something like `printf("(");` the outer
closing parenthesis matches with the opening parenthesis inside of the
quotes rather than the opening parenthesis starting your function call.
This can also mess up parentheses matching in following lines.

There is the warning of the performance hit but I find that breaking
matching parentheses makes it almost unusable so I'd rather take the
performance hit. If the performance hit is noticeable or really bad I
may look for alternative solutions.